refactor(storage): Split columnar.rs into submodules
Convert single-file columnar module (1130 lines) into a module directory
with focused submodules:

- types.rs: ColumnTypeClass enum and type classification
- data.rs: ColumnData enum with typed storage and NULL bitmap
- builder.rs: ColumnBuilder for pre-allocated column construction
- table.rs: ColumnarTable with row/columnar conversion
- mod.rs: Public API re-exports and documentation

Maintains all public APIs (ColumnData, ColumnarTable) and all existing
tests pass. No functional changes.

Closes #2703
